New allocator from BECS added

+++ Editor/Utils/WorldEditor.cs
@@ -43,7 +43,7 @@ namespace ME.ECSEditor {
 
                     try {
 
-                        this.stateSize = this.world.GetState().allocator.GetReservedSize();
+                        this.stateSize = (int)this.world.GetState().allocator.GetReservedSize();
                         this.stateSize += UnityObjectUtils.GetObjectSize(this.world.GetState());
 
                     } catch (System.Exception ex) {
